Move mac.go to lib/crypto
